requirements.txt
pandas==1.5.3
pytest==7.2.1
matplotlib==3.6.3
python-dotenv==0.21.1
scikit-learn==1.2.0